requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
streamlit>=1.45.1
python-dotenv>=1.1.0
chromadb>=1.0.9
llama-index>=0.12.36
llama-index-vector-stores-chroma>=0.4.1
llama-index-embeddings-huggingface>=0.5.4
sentence-transformers>=4.1.0
pillow>=11.1.0
openai>=1.79.0